refactor(chat): update folder structure
