+++ src/renderer/src/app.tsx
@@ -7,7 +7,7 @@ import type { FC } from "react";
 import { Navigate, Route, Routes } from "react-router-dom";
 
 import { AgentsPage } from "@features/agents/agents-page";
-import { ChatPage } from "@features/chat/chat-page";
+import { ChatPage } from "@features/chat/components/chat-page";
 import { OnboardingModal } from "@features/onboarding";
 import { SettingsPage } from "@features/settings/settings-page";
 import { SidebarNavigation } from "@renderer/components/navigation/sidebar-navigation";
