+++ src/renderer/src/features/chat/components/canvas/canvas-tabs.tsx
