+++ src/renderer/src/features/chat/components/canvas/canvas-toolbar.tsx
