+++ src/renderer/src/features/chat/components/canvas/code-editor.tsx
