+++ src/renderer/src/features/chat/components/canvas/index.tsx
