+++ src/renderer/src/features/chat/components/canvas/types.ts
