+++ src/renderer/src/features/chat/components/chat-content.tsx
@@ -13,11 +13,11 @@ import { Canvas } from "./canvas";
 import { ChatHeader } from "./chat-header";
 import { ChatOptionsSidebar } from "./chat-options-sidebar";
 import { ChatTabs } from "./chat-tabs";
-import { ChatUtilities } from "./chat-utilities";
+import { ChatUtilities } from "../utils/chat-utilities";
 import { MessageInput } from "./message-input";
 import { MessagesView } from "./messages-view";
 import { RawInfoView } from "./raw-info-view";
-import type { Message } from "./types";
+import type { Message } from "../types";
 
 const DEFAULT_MESSAGE_SUGGESTIONS = [
 	"Go to my documents folder",
