+++ src/renderer/src/features/chat/components/chat-options-sidebar-styled.tsx
