+++ src/renderer/src/features/chat/components/chat-sidebar.tsx
