+++ src/renderer/src/features/chat/components/chat-tabs.tsx
