+++ src/renderer/src/features/chat/components/directory-indicator.tsx
