+++ src/renderer/src/features/chat/components/error-view.tsx
