+++ src/renderer/src/features/chat/components/loading-indicator.tsx
