+++ src/renderer/src/features/chat/components/message-item/action-highlight.tsx
