+++ src/renderer/src/features/chat/components/message-item/collapsible-message.tsx
