+++ src/renderer/src/features/chat/components/message-item/collapsible-section.tsx
