+++ src/renderer/src/features/chat/components/message-item/error-block.tsx
