+++ src/renderer/src/features/chat/components/message-item/image-attachment.tsx
