+++ src/renderer/src/features/chat/components/message-item/invalid-attachment.tsx
