+++ src/renderer/src/features/chat/components/message-item/log-block.tsx
