+++ src/renderer/src/features/chat/components/message-item/message-avatar.tsx
