+++ src/renderer/src/features/chat/components/message-item/message-paper.tsx
@@ -1,8 +1,8 @@
 import { Box, Paper, useTheme } from "@mui/material";
 import { styled } from "@mui/material/styles";
 import { type FC, useEffect, useMemo, useRef } from "react";
-import { useScrollToBottom } from "../../../hooks/use-scroll-to-bottom";
-import type { Message } from "../types";
+import { useScrollToBottom } from "../../../../hooks/use-scroll-to-bottom";
+import type { Message } from "../../types";
 import { MessageControls } from "./message-controls";
 import { MessageTimestamp } from "./message-timestamp";
 import { StreamingMessage } from "./streaming-message";
