+++ src/renderer/src/features/chat/components/message-item/output-block.tsx
