+++ src/renderer/src/features/chat/components/message-item/security-check-highlight.tsx
