+++ src/renderer/src/features/chat/components/message-item/status-indicator.tsx
