+++ src/renderer/src/features/chat/components/message-item/streaming-message.tsx
@@ -5,9 +5,9 @@ import { Box, CircularProgress, Typography, styled } from "@mui/material";
  * @module StreamingMessage
  */
 import { useCallback, useEffect, useMemo, useRef } from "react";
-import type { AgentExecutionRecord } from "../../../api/local-operator/types";
-import { useStreamingMessage } from "../../../hooks/use-streaming-message";
-import { useStreamingMessagesStore } from "../../../store/streaming-messages-store";
+import type { AgentExecutionRecord } from "../../../../api/local-operator/types";
+import { useStreamingMessage } from "../../../../hooks/use-streaming-message";
+import { useStreamingMessagesStore } from "../../../../store/streaming-messages-store";
 import { MessageContent } from "./message-content";
 
 // Styled components
