+++ src/renderer/src/features/chat/components/placeholder-view.tsx
