+++ src/renderer/src/features/chat/components/scroll-to-bottom-button.tsx
