+++ src/renderer/src/features/chat/components/unset-slider-setting.tsx
