+++ src/renderer/src/features/chat/utils/chat-options-utils.tsx
